// src/app/models/userupdate.ts
export interface userupda{

    name?: string;
    Sname?: string;
    lastName?:string;
    SlastName?:string;
    email?: string; 
    phone ?: number
    password?: string,
    rol ?: string , 
    subjectIds?: number[];
    courseIds?: number[];


}

export interface status {
  isActive: boolean; 
}